Adds "stage" to bastion name for better uniqueness;
this allows us to deploy 2+ bastions in a single account without
having to worry about name collisions. As S3 uses a global namespace
we cannot have 2 buckets with identical names.
This will help better identify which server is used for what environment
when multiple environments are put into a single account/region.